# ARROW-2354: [C++] Make PyDecimal_Check() faster