This basically keeps an eternal reference to the decimal type.

Before:
```
[100.00%] ··· Running convert_pandas.PandasConversionsToArrow.time_from_series                                                                               ok
[100.00%] ···· 
               ========= ========== ============= ============== ===========
               --                               dtype                       
               --------- ---------------------------------------------------
                  size     int64       float64     float64_nans      str    
               ========= ========== ============= ============== ===========
                   10     420±1μs      426±1μs      421±0.3μs     450±0.6μs 
                1000000   6.92±1ms   14.1±0.02ms   15.5±0.04ms      1.66s   
               ========= ========== ============= ============== ===========
```

After:
```
[100.00%] ··· Running convert_pandas.PandasConversionsToArrow.time_from_series                                                                               ok
[100.00%] ···· 
               ========= =========== ============= ============== ===========
               --                               dtype                        
               --------- ----------------------------------------------------
                  size      int64       float64     float64_nans      str    
               ========= =========== ============= ============== ===========
                   10     425±0.9μs     428±1μs      430±0.8μs     438±0.6μs 
                1000000    7.49±1ms   14.2±0.04ms   15.4±0.05ms     99.2±1ms 
               ========= =========== ============= ============== ===========
```