@wesm 2018-03-26
Can you add an ASV benchmark to exercise the performance issue in https://github.com/apache/arrow/issues/1792?