@pitrou 2018-03-26
The benchmark already exists, see above :-)