@wesm 2018-03-26
@cpcloud I'll wait for you to have a last look at this before merging (you have a comment unaddressed still)