@cpcloud 2018-03-26
I'll add the patch and bump the build number to `arrow-cpp`.