@cpcloud 2018-03-26
@pitrou @wesm Follow-up JIRA'd: https://issues.apache.org/jira/browse/ARROW-2357